fix(base-unit): loi LATENT nghiem trong quet nhap kho (hut ton x units_per_carton) + dong bo hien thi thung+hop Stocktake/PalletOps/Inbound pane
CRITICAL - InboundScanSheet: o so luong dung Input tho gui cartons_override = cartons_per_pallet (THUNG VAT LY) lam BASE -> ma co entry ghi ton nhap HUT x units_per_carton o lan quet QR nhap sau flip (data hien tai pre-flip nen chua manifest). Fix: default = effCartonsPerPallet x units_per_carton (base) + doi Input -> QtyInput 2 o Thung+Hop (quy doi tai ria nhu moi sheet quet). Go import Input thua.

Display base tho gan nhan thung -> qtyLabel/qtyEntryText:
- StocktakeDashboard: DetailPanel (Ton thuc te/app/Chenh lech), cot bang Thuc te/Chenh lech (khop qtyEntryText cot Ton App), export Excel qtyEntryDecimal
- PalletOps: msg tach 'goc con {base} thung' + SummaryBand tab Tach (Ton goc/Tach ra/Giu lai) -> qtyLabel
- Inbound pane phai: tile 'Thuc nhap' base tho -> qtyEntryText + don vi
- Stocktake: QtyInput value parseInt -> parseFloat (ma KG nhap le duoc thap phan)

+++ frontend/src/pages/wms/Stocktake.tsx
@@ -365,7 +365,7 @@ export default function Stocktake() {
                     </p>
                     <div className="flex items-center gap-2">
                       <QtyInput compact autoFocus className="w-44"
-                        value={Math.max(0, parseInt(physCount) || 0)}
+                        value={Math.max(0, parseFloat(physCount) || 0)}
                         mat={entry.material}
                         onChange={b => setPhysCount(String(b))}
                       />
